% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-diff.R
\name{col_diff}
\alias{col_diff}
\title{Strip column from strings contained in another column.}
\usage{
col_diff(df, left, right, how = "exact")
}
\arguments{
\item{df}{A data frame object or something coercible to one.}

\item{left}{Column object to transform, following tidyverse conventions.
Column to remove from, (left - right).}

\item{right}{Column object to transform, following tidyverse conventions.
Content to remove (left - right).}

\item{how}{exact, lowercase, squish, token, levenshtein. Can be a combination
such as c('lowercase', 'squish').}
}
\value{
A data frame object.
}
\description{
Strip column from strings contained in another column.
}
\examples{
col_diff(businesses[1:8, 2:3],
  address,
  postcode,
  how = c("lowercase", "squish")
)
}
