% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco_us}
\alias{eco_us}
\title{Ecoregions of the United States}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 330 rows and 5 columns.
}
\source{
https://www.sciencebase.gov/catalog/item/54244abde4b037b608f9e23d
}
\usage{
eco_us
}
\description{
Bailey's Ecoregions of the Conterminous United States
}
\keyword{datasets}
