% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape-string.R
\name{escape_string}
\alias{escape_string}
\title{Escape characters in string. Make safe for string operations.}
\usage{
escape_string(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector.
}
\description{
Escape characters in string. Make safe for string operations.
}
\examples{
escape_string("(Unit 8) 23 Woodland Road")
}
