% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{find_buildings_numbers}
\alias{find_buildings_numbers}
\title{Find the numerical parts (building|house|unit) in a list, vector or data
frame of addresses.}
\usage{
find_buildings_numbers(X, ...)
}
\arguments{
\item{X}{list, vector or data frame, appropriate to a call to lapply.}

\item{...}{optional arguments to \code{find_buildings_numbers_in_string()}.}
}
\value{
list
}
\description{
Returns an empty string where no match is found.
Vectorized version of \code{find_buildings_numbers_in_string()}.Doest not match
postcodes or parts of a postcode.
}
\examples{
find_buildings_numbers(businesses)
}
\seealso{
\code{\link[=find_buildings_numbers_in_string]{find_buildings_numbers_in_string()}}
}
