% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{find_buildings_numbers_in_string}
\alias{find_buildings_numbers_in_string}
\title{Find the numerical parts of an address (building|house|unit).}
\usage{
find_buildings_numbers_in_string(string, locale = "GBR", sep = "|")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one (e.g. a data frame).}

\item{locale}{A string, the country code in format ISO 3. Default is "GBR".}

\item{sep}{A character string to separate the terms in the returned string.}
}
\value{
A character vector of the contatenated values (shape: (1,1)).
}
\description{
Returns an empty string where no match is found.
The match is performed with a look ahead to match address number patterns
that are not a postcode.
For large flat files, use \code{find_buildings_numbers()}.
}
\examples{
string <- "The quick brown fox's family lives at 22A-22B Bridge Road, N17 0RN."
find_buildings_numbers_in_string(string)

}
\seealso{
\code{\link[=find_buildings_numbers]{find_buildings_numbers()}}
}
