% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{find_postcodes}
\alias{find_postcodes}
\title{Find postcodes in a list, vector or data frame of strings/addresses.}
\usage{
find_postcodes(X, ...)
}
\arguments{
\item{X}{list, vector or data frame, appropriate to a call to lapply}

\item{...}{optional arguments to \code{find_postcodes_in_string()}.}
}
\value{
list
}
\description{
Returns an empty string where no match is found.
Vectorized version of \code{find_postcodes_in_string()}.
}
\examples{
find_postcodes(businesses)
}
\seealso{
\code{\link[=find_postcodes_in_string]{find_postcodes_in_string()}}
}
