% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{format_postcode}
\alias{format_postcode}
\title{Reformat a postcode to standards.}
\usage{
format_postcode(postcode, locale = "GBR")
}
\arguments{
\item{postcode}{A character string or vector of strings.}

\item{locale}{A string, the country code in format ISO 3. Default is "GBR".}
}
\value{
A character string.
}
\description{
For GBR postcodes, if it is missing a white space between the outward
(area + district) and inward (sector + unit) codes, then add it.
}
\examples{
format_postcode('eC2A0Rn')

}
