% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gapfill_postcodes}
\alias{gapfill_postcodes}
\title{gapfill_postcodes}
\usage{
gapfill_postcodes(postcode, address)
}
\arguments{
\item{postcode}{a vector of strings (same size as address)}

\item{address}{a vector of strings (same size as postcode)}
}
\description{
Find postcodes in address vector when missing or incomplete in postcode vector.
}
