% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{is_postcode_complete}
\alias{is_postcode_complete}
\title{Is the string a full postcode.}
\usage{
is_postcode_complete(postcode, ignore_case = FALSE)
}
\arguments{
\item{postcode}{A character string.}

\item{ignore_case}{A boolean.}
}
\value{
A boolean.
}
\description{
Is the string a full postcode.
}
\section{TODO}{

Handle case is_postcode_complete("ng22 9qc", ignore_case = TRUE)
This postcode is not recognized by the regex but is valid(?)
}

\examples{
is_postcode_complete("EC2A 3JX")
is_postcode_complete("ec2a", ignore_case = TRUE)
}
