% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{is_postcode_partial}
\alias{is_postcode_partial}
\title{Is the string a partial postcode.}
\usage{
is_postcode_partial(postcode, ignore_case = FALSE)
}
\arguments{
\item{postcode}{A character string.}

\item{ignore_case}{A boolean.}
}
\value{
A boolean.
}
\description{
A postcode is partial if:
\enumerate{
\item it is not complete
\item it is the area + district (outward) part of the postcode
\item OR it is the sector + unit (inward) part of the postcode
}
}
\examples{
is_postcode_partial("EC2A 3JX")
is_postcode_partial("ec2a", ignore_case = TRUE)
}
