% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.R
\name{melt_rows}
\alias{melt_rows}
\title{Separate a collapsed column record into multiple rows.}
\usage{
melt_rows(df, column, dividers = c("\\\\|"), trim_ws = TRUE)
}
\arguments{
\item{df}{A \code{DataFrame}.}

\item{column}{The name of the column/feature to perfom the separation on.}

\item{dividers}{A vector of delimiters to}

\item{trim_ws}{A boolean; determines if trailing white spaces (potentially
generated by the split), are trimmed or not.}
}
\value{
A \code{DataFrame} with the additional rows (as many as performed splits).
}
\description{
Wrapper around \code{tidyr::separate_rows()}. Performs the operation for a given
list of dividers (delimiters).
Iterates through the set of possible dividers.
}
\details{
If a feature/variable in a record contains multiple delimited values, this
separates the values, placing each on in its own row (creating a new record).
The transformation is similar to wide to long, but only (feature, record)
having delimited values will create new rows.
}
\examples{
melt_rows(
  businesses[4:8, 3:4],
  "trading_name",
  dividers = c("\\\\|", " trading as ", "t/a", "\\\\/")
  )

}
