% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_nested_records}
\alias{parse_nested_records}
\title{parse_nested_records}
\usage{
parse_nested_records(s, sep_from, sep_to)
}
\arguments{
\item{s}{a vector of strings}

\item{sep_from}{string separator in input strings}

\item{sep_to}{string separator in output strings}
}
\description{
Deduplicate and optionally change separators in nested records.
}
