% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partitions2}
\alias{partitions2}
\title{Blue regionalization (raster version)}
\format{
An object of class \code{RasterLayer} of dimension 8 x 10 x 1.
}
\usage{
partitions2
}
\description{
Raster data of the blue regionalization used in Figure 1 of Stepinski and Nowosad (2018)
}
\references{
Nowosad, Jakub, and Tomasz F. Stepinski.
"Spatial association between regionalizations using the information-theoretical V-measure."
International Journal of Geographical Information Science (2018).
https://doi.org/10.1080/13658816.2018.1511794
}
\keyword{datasets}
