% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions1}
\alias{regions1}
\title{Red regionalization}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 2 columns.}
\usage{
regions1
}
\description{
Data of the red regionalization used in Figure 1 of Stepinski and Nowosad (2018)
}
\references{
Stepinski, Tomasz, and Jakub Nowosad. "Assessing a degree of
spatial association between regionalizations or categorical maps using the
information-theoretical V-measure."
}
\keyword{datasets}
