% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_prep.R
\name{regions_prep}
\alias{regions_prep}
\title{Regionalizations preparation}
\usage{
regions_prep(x, y, x_name, y_name)
}
\arguments{
\item{x}{DESC}

\item{y}{DESC}

\item{x_name}{DESC}

\item{y_name}{DESC}
}
\value{
A list
}
\description{
The output of this function can be used in \code{\link[=sabre_calc]{sabre_calc()}} and \code{\link[=mapcurves_calc]{mapcurves_calc()}}.
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
