% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sabre-data}
\alias{sabre-data}
\title{Sample tibble vectors for practicing data cleaning and running tests}
\format{
tibble
}
\description{
\code{businesses} is a collection of business records with 4 features
"index":
"address":
"postcode":
"trading_name":
}
\examples{
find_buildings_numbers(businesses)
}
