% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabre_calc.R
\name{sabre_calc}
\alias{sabre_calc}
\title{SABRE calculation}
\usage{
sabre_calc(x, x_name, y, y_name, B = 1, precision = 1)
}
\arguments{
\item{x}{DESC}

\item{x_name}{DESC}

\item{y}{DESC}

\item{y_name}{DESC}

\item{B}{DESC}

\item{precision}{DESC}
}
\value{
A list
}
\description{
It calculate the Spatial Association Between REgionalizations (SABRE).
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Rosenberg, Andrew, and Julia Hirschberg. "V-measure:
A conditional entropy-based external cluster evaluation measure." Proceedings
of the 2007 joint conference on empirical methods in natural language
processing and computational natural language learning (EMNLP-CoNLL). 2007.
}
