% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{standardize_strings}
\alias{standardize_strings}
\title{Standardize the character columns of a \code{DataFrame} object.}
\usage{
standardize_strings(df, columns, replacements = "ghbusiness")
}
\arguments{
\item{df}{A \code{DataFrame} object.}

\item{columns}{A character vector of the columns' names to perform the
operation on.}

\item{replacements}{The replacements to operate in the string.

A list of element-wise mapping of patterns to be replaced with their
replacements.
A the name of a pre-defined set of replacements, contained in the
replacements_ environmental list object variable.}
}
\value{
A \code{DataFrame} object
}
\description{
Given a \code{DataFrame}, a set of columns and a list of replacements,
will perform a standard pre-processing of strings: a) lowercasing, b)
characters replacements, c) white spaces squishing (removes whitespace from
start and end of string, also reduces repeated whitespace inside a string).
}
\examples{
columns <- c("address", "trading_name")
standardize_strings(businesses, columns, replacements='ghbusiness')

columns <- c("address", "trading_name")
from <- c(',', '\\\\(.*\\\\)', '&')
to <- c('', '', ' and ')
standardize_strings(
  businesses,
  columns,
  replacements=list(from, to)
  )
}
