% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{strip_buildings_numbers}
\alias{strip_buildings_numbers}
\title{Strip buildings numbers (building|house|unit) from a list, vector or data
frame of addresses.}
\usage{
strip_buildings_numbers(X, ...)
}
\arguments{
\item{X}{list, vector or data frame, appropriate to a call to lapply}

\item{...}{optional arguments to \code{find_buildings_numbers_in_string()}.}
}
\value{
list
}
\description{
Strip buildings numbers (building|house|unit) from a list, vector or data
frame of addresses.
}
\examples{
strip_buildings_numbers(businesses)
}
\seealso{
\code{\link[=strip_buildings_numbers_in_string]{strip_buildings_numbers_in_string()}}
}
