% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{strip_buildings_numbers_in_string}
\alias{strip_buildings_numbers_in_string}
\title{Strip buildings numbers (building|house|unit)}
\usage{
strip_buildings_numbers_in_string(string, locale = "GBR")
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one (e.g. a data frame).}

\item{locale}{A string, the country code in format ISO 3. Default is "GBR".}
}
\value{
A character vector of the original character strings striped from
building numbers.
}
\description{
Strip buildings numbers (building|house|unit)
}
\examples{
strip_buildings_numbers_in_string(businesses[['address']])

}
\seealso{
\code{\link[=strip_buildings_numbers]{strip_buildings_numbers()}}
}
