% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{strip_buildings_numbers_list}
\alias{strip_buildings_numbers_list}
\title{Strip buildings numbers (building|house|unit) from a list, or data
frame of addresses.}
\usage{
strip_buildings_numbers_list(X, ...)
}
\arguments{
\item{X}{list or data frame, appropriate for a call to lapply}

\item{...}{optional arguments to \code{strip_buildings_numbers_vec()}.}
}
\value{
list
}
\description{
Strip buildings numbers (building|house|unit) from a list, or data
frame of addresses.
}
\examples{
strip_buildings_numbers_list(businesses)
}
\seealso{
\code{\link[=strip_buildings_numbers_vec]{strip_buildings_numbers_vec()}}
}
