% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{strip_buildings_numbers_vec}
\alias{strip_buildings_numbers_vec}
\title{Strip buildings numbers (building|house|unit) from a string or
character vector of addresses.}
\usage{
strip_buildings_numbers_vec(string, locale = "GBR", squish_ws = TRUE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{locale}{A string, the country code in format ISO 3. Default is "GBR".}

\item{squish_ws}{A boolean; determines if trailing white spaces and repeated
whitespace inside a string (generated by the replacement), are trimmed or
not.}
}
\value{
A character vector of the original character strings striped from
building numbers.
}
\description{
Strip buildings numbers (building|house|unit) from a string or
character vector of addresses.
}
\examples{
strip_buildings_numbers_vec(businesses[['address']])

}
\seealso{
\code{\link[=strip_buildings_numbers_vec]{strip_buildings_numbers_vec()}}
}
