% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{strip_business_legal_entity_type}
\alias{strip_business_legal_entity_type}
\title{Strip business legal entity type from name}
\usage{
strip_business_legal_entity_type(string, trim_ws = FALSE, squish_ws = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{trim_ws}{A boolean; determines if trailing white spaces (generated by
the replacement), are trimmed or not.}

\item{squish_ws}{A boolean; determines if trailing white spaces and repeated
whitespace inside a string (generated by the replacement), are trimmed or
not.}
}
\value{
A character vector.
}
\description{
Strip business legal entity type from name
}
\examples{
strip_business_legal_entity_type("Farmer Box Plc.")

}
