% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_measure.R
\name{v_measure}
\alias{v_measure}
\title{V-measure}
\usage{
v_measure(x, y, z = NULL, B = 1)
}
\arguments{
\item{x}{A numeric vector of counts}

\item{y}{A numeric vector of counts}

\item{z}{A numeric matrix of counts}

\item{B}{DESC}
}
\value{
A list
}
\description{
A conditional entropy-based external cluster evaluation measure.
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Rosenberg, Andrew, and Julia Hirschberg. "V-measure:
A conditional entropy-based external cluster evaluation measure." Proceedings
of the 2007 joint conference on empirical methods in natural language
processing and computational natural language learning (EMNLP-CoNLL). 2007.
}
