% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{find_buildings_numbers}
\alias{find_buildings_numbers}
\title{Find the numerical parts (building|house|unit) in a list, vector or data
frame of addresses.}
\usage{
find_buildings_numbers(X)
}
\arguments{
\item{X}{list, vector or data frame, appropriate to a call to sapply.}
}
\value{
A named vector of characters.
}
\description{
Find the numerical parts (building|house|unit) in a list, vector or data
frame of addresses.
}
\examples{
find_buildings_numbers(businesses[['address']])

}
