% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcurves.R
\name{mapcurves}
\alias{mapcurves}
\title{Mapcurves}
\usage{
mapcurves(x, y, z = NULL)
}
\arguments{
\item{x}{A numeric vector of counts}

\item{y}{A numeric vector of counts}

\item{z}{A numeric matrix of counts}
}
\value{
A list
}
\description{
Mapcurves: a quantitative method for comparing categorical maps.
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Hargrove, William W., Forrest M. Hoffman, and Paul F. Hessburg.
"Mapcurves: a quantitative method for comparing categorical maps."
Journal of Geographical Systems 8.2 (2006): 187.
}
