% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcurves_calc.R
\name{mapcurves_calc}
\alias{mapcurves_calc}
\title{Mapcurves calculation}
\usage{
mapcurves_calc(x)
}
\arguments{
\item{x}{An output of the \code{\link[=regions_prep]{regions_prep()}} function}
}
\value{
A list
}
\description{
It calculate the Mapcurves's goodness-of-fit (GOF) using
the output of the \code{\link[=regions_prep]{regions_prep()}} function.
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Hargrove, William W., Forrest M. Hoffman, and Paul F. Hessburg.
"Mapcurves: a quantitative method for comparing categorical maps."
Journal of Geographical Systems 8.2 (2006): 187.
}
