% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcurves_calc.R
\name{mapcurves_calc}
\alias{mapcurves_calc}
\title{Mapcurves calculation}
\usage{
mapcurves_calc(x, x_name, y, y_name, precision = 1)
}
\arguments{
\item{x}{DESC}

\item{x_name}{DESC}

\item{y}{DESC}

\item{y_name}{DESC}

\item{precision}{DESC}
}
\value{
A list
}
\description{
It calculates the Mapcurves's goodness-of-fit (GOF)
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Hargrove, William W., Forrest M. Hoffman, and Paul F. Hessburg.
"Mapcurves: a quantitative method for comparing categorical maps."
Journal of Geographical Systems 8.2 (2006): 187.
}
