% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions1}
\alias{regions1}
\title{Red regionalization}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
regions1
}
\description{
Data of the red regionalization used in Figure 1 of Stepinski and Nowosad (2018)
}
\references{
Nowosad, Jakub, and Tomasz F. Stepinski.
"Spatial association between regionalizations using the information-theoretical V-measure."
International Journal of Geographical Information Science (2018).
https://doi.org/10.1080/13658816.2018.1511794
}
\keyword{datasets}
