% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabre-package.R
\docType{package}
\name{sabre-package}
\alias{sabre}
\alias{sabre-package}
\title{sabre: Spatial Association Between REgionalizations}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

Authors:
\itemize{
  \item Tomasz Stepinski
}

Other contributors:
\itemize{
  \item Space Informatics Lab [copyright holder]
}

}
\keyword{internal}
