% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabre_calc.R
\name{sabre_calc}
\alias{sabre_calc}
\title{SABRE calculation}
\usage{
sabre_calc(x, x_name, y, y_name, B = 1, precision = NULL)
}
\arguments{
\item{x}{An object of class \code{sf}with a \code{POLYGON} or \code{MULTIPOLYGON} geometry type}

\item{x_name}{A name of the column with regions/clusters names.}

\item{y}{An object of class \code{sf}with a \code{POLYGON} or \code{MULTIPOLYGON} geometry type}

\item{y_name}{A name of the column with regions/clusters names.}

\item{B}{A numeric value. If \code{B`` > 1 then completeness is weighted more strongly than homogeneity, and if}B` < 1 then homogeneity is weighted more strongly than
completeness}

\item{precision}{A numeric value. DESC}
}
\value{
A list
}
\description{
It calculate the Spatial Association Between REgionalizations (SABRE).
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
...

Rosenberg, Andrew, and Julia Hirschberg. "V-measure:
A conditional entropy-based external cluster evaluation measure." Proceedings
of the 2007 joint conference on empirical methods in natural language
processing and computational natural language learning (EMNLP-CoNLL). 2007.
}
