% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sabre_calc.R
\name{sabre_calc}
\alias{sabre_calc}
\title{SABRE calculation}
\usage{
sabre_calc(x, unit = "log2", B = 1)
}
\arguments{
\item{x}{An output of the \code{\link[=regions_prep]{regions_prep()}} function}

\item{unit}{A logarithm base ("log", "log2" or "log10")}

\item{B}{DESC}
}
\value{
A list
}
\description{
It calculate the Spatial Association Between REgionalizations (SABRE) using
the output of the \code{\link[=regions_prep]{regions_prep()}} function.
}
\details{
DETAILS
}
\examples{
# EXAMPLES

}
\references{
Rosenberg, Andrew, and Julia Hirschberg. "V-measure:
A conditional entropy-based external cluster evaluation measure." Proceedings
of the 2007 joint conference on empirical methods in natural language
processing and computational natural language learning (EMNLP-CoNLL). 2007.
}
