% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{strip_buildings_numbers}
\alias{strip_buildings_numbers}
\title{Strip buildings numbers (building|house|unit).}
\usage{
strip_buildings_numbers(.data, locale = "GBR", squish_ws = TRUE)
}
\arguments{
\item{.data}{chacter, character vector, list or data frame.}

\item{locale}{A string, the country code in format ISO 3. Default is "GBR".}

\item{squish_ws}{A boolean; determines if trailing white spaces and repeated
whitespace inside a string (generated by the replacement), are trimmed or
not.}
}
\value{
An object of the same type as .data.
}
\description{
Strip buildings numbers (building|house|unit).
}
\examples{
strip_buildings_numbers(businesses[['address']])
strip_buildings_numbers(businesses)

}
